Add react routing to app
