+++ ui/package.json
@@ -11,10 +11,12 @@
     "@types/node": "^12.12.26",
     "@types/react": "^16.9.19",
     "@types/react-dom": "^16.9.5",
+    "@types/react-router-dom": "^5.1.3",
     "axios": "^0.19.2",
     "dom-to-image-more": "^2.8.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.1",
     "typescript": "^3.7.5"
   },
