+++ ui/src/App.test.tsx
@@ -1,9 +1,15 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import App from './App';
+import { createMemoryHistory } from "history";
+import { Router } from "react-router";
 
-test('renders page title', () => {
-  const { getByText } = render(<App />);
-  const headerElement = getByText(/create your meme/i);
-  expect(headerElement).toBeInTheDocument();
+test('renders landing page', () => {
+  const history = createMemoryHistory();
+  render(
+    <Router history={history}>
+      <App />
+    </Router>
+  );
+  expect(history.location.pathname).toBe("/");
 });
