+++ ui/src/App.tsx
@@ -1,25 +1,17 @@
-import React, { useReducer } from 'react';
-import { AppDispatchContext, AppDispatchReducer, initialState } from './reducers/AppDispatchReducer';
-import './App.css';
-import MemeUpload from './MemeUpload';
-import ImageUpload from './ImageUpload';
+import React from 'react';
+import {Route, Switch, BrowserRouter as Router } from "react-router-dom";
+import Upload from './CreateMeme';
+import RouterError from './RouterError';
 
 function App() {
-  const [appState, dispatch] = useReducer(AppDispatchReducer, initialState);
-  const value = {state: appState, dispatch};
 
   return (
-    <div className="App">
-      <header className="App-header">
-        <h1 className="inputPrompt">Create Your Meme</h1>
-        <div className="memeCreation">
-          <AppDispatchContext.Provider value={value}>
-            <ImageUpload topText={appState.topText} bottomText={appState.bottomText} memeContentRef={appState.memeContentRef}/>
-            <MemeUpload memeName={appState.memeName} imageID={appState.imageID} memeContentRef={appState.memeContentRef} memeResultRef={appState.memeResultRef}/>
-          </AppDispatchContext.Provider>
-        </div>
-      </header>
-    </div>
+    <Router>
+      <Switch>
+        <Route path = "/" component = {Upload} exact />
+        <Route component = {RouterError} />
+      </Switch>
+    </Router>
   );
 
 }
