+++ ui/src/CreateMeme.css
@@ -1,0 +1,19 @@
+.Create {
+    text-align: center;
+}
+  
+.Create-header {
+    background-color: #282c34;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+}
+  
+.memeCreation {
+    display: flex;
+    align-items: center;
+}
